feat: refactor release note handling by adding utility functions for body retrieval and dependency wrapping
